requirements.txt
fastapi
uvicorn
twitchio
python-dotenv
websockets
requests
